Do not force prefix to C-a
Assuming no one uses C-b as the default prefix is a wrong assumption.

Instead of making any assumption about the prefix, we're "building" on
top of whatever the user has set.

The following key-bindings will be set if prefix is set to C-a:
- `C-a + C-a` - sends prefix character
- `C-a + a` - switches to the last window

Likewise, if C-b is prefix, then the additional bindings will be
`C-b + C-b` and `C-b + b`.

Closes #1